+++ src/stock-trading-vue/vite.config.ts
@@ -1,10 +1,21 @@
 import { fileURLToPath, URL } from 'node:url'
 
-import { defineConfig } from 'vite'
+import { defineConfig, loadEnv } from 'vite'
 import vue from '@vitejs/plugin-vue'
+
+const env = loadEnv(process.env.NODE_MODE, process.cwd())
 
 // https://vitejs.dev/config/
 export default defineConfig({
+  server: {
+    proxy: {
+      '/stock_trading_api': {
+        target: env.VITE_API_URL + '/' + env.VITE_API_VERSION,
+        changeOrigin: true,
+        rewrite: (path) => path.replace(/^\/stock_trading_api/, '')
+      }
+    }
+  },
   plugins: [
     vue(),
   ],
@@ -12,5 +23,5 @@ export default defineConfig({
     alias: {
       '@': fileURLToPath(new URL('./static', import.meta.url))
     }
-  }
+  },
 })